Added Login Validation to Routes

+++ LoginContacts/Controllers/UserController.cs
@@ -8,11 +8,10 @@ using Microsoft.EntityFrameworkCore;
 using Login.Models;
 using LoginContacts.Data;
 using System.Diagnostics;
+using Microsoft.AspNetCore.Http;
 
-namespace LoginContacts.Controllers
-{
-    public class UserController : Controller
-    {
+namespace LoginContacts.Controllers {
+    public class UserController : Controller {
         private readonly LoginContactsContext _context;
 
         public UserController(LoginContactsContext context)
@@ -23,12 +22,23 @@ namespace LoginContacts.Controllers
         // GET: User
         public async Task<IActionResult> Index()
         {
-            return View(await _context.User.ToListAsync());
+            Console.WriteLine("current session: " + HttpContext.Session.GetString("_UserSession"));
+            if(UserIsLoggedIn()) {
+                return View(await _context.User.ToListAsync());
+            }
+
+            return RedirectToAction(nameof(Login));
         }
 
         public async Task<IActionResult> Login()
         {
             return View();
+        }
+
+        public async Task<IActionResult> Logout()
+        {
+            HttpContext.Session.Remove("_UserSession");
+            return RedirectToAction(nameof(Login));
         }
 
         [HttpPost]
@@ -38,10 +48,12 @@ namespace LoginContacts.Controllers
                 .FirstOrDefaultAsync(m => 
                     m.Email == user.Email && m.Password == user.Password
                 );
+
             if (usr == null) {
                 Console.WriteLine("No user found");
             } else {
-               return RedirectToAction("Details", new {id = usr.Id});
+                HttpContext.Session.SetString("_UserSession", usr.Id.ToString());
+                return RedirectToAction("Details", new {id = usr.Id});
             }
             
             return View();
@@ -50,21 +62,25 @@ namespace LoginContacts.Controllers
         // GET: User/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
+            if(UserIsLoggedIn()) {
+                if (id == null)
+                {
+                    return NotFound();
+                }
+
+                var user = await _context.User
+                    .Include(u => u.Contacts) 
+                    .FirstOrDefaultAsync(m => m.Id == id);
+
+                if (user == null)
+                {
+                    return NotFound();
+                }
+
+                return View(user);
             }
 
-            var user = await _context.User
-                .Include(u => u.Contacts) 
-                .FirstOrDefaultAsync(m => m.Id == id);
-
-            if (user == null)
-            {
-                return NotFound();
-            }
-
-            return View(user);
+            return RedirectToAction(nameof(Login));
         }
 
         // GET: User/Create
@@ -106,21 +122,25 @@ namespace LoginContacts.Controllers
         // GET: User/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            if(UserIsLoggedIn()) {
+                if (id == null)
+                {
+                    return NotFound();
+                }
 
-            var user = await _context.User
-                .Include(u => u.Contacts) 
-                .FirstOrDefaultAsync(m => m.Id == id);
+                var user = await _context.User
+                    .Include(u => u.Contacts) 
+                    .FirstOrDefaultAsync(m => m.Id == id);
 
-            if (user == null)
-            {
-                return NotFound();
+                if (user == null)
+                {
+                    return NotFound();
+                }
+                
+                return View(user);
             }
             
-            return View(user);
+            return RedirectToAction(nameof(Login));
         }
 
         // POST: User/Edit/5
@@ -130,21 +150,76 @@ namespace LoginContacts.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("Id,Email,Password")] User user)
         {
-            if (id != user.Id)
-            {
-                return NotFound();
+            if(UserIsLoggedIn()) {
+                if (id != user.Id)
+                {
+                    return NotFound();
+                }
+
+                if (ModelState.IsValid)
+                {
+                    try
+                    {
+                        _context.Update(user);
+                        await _context.SaveChangesAsync();
+                    }
+                    catch (DbUpdateConcurrencyException)
+                    {
+                        if (!UserExists(user.Id))
+                        {
+                            return NotFound();
+                        }
+                        else
+                        {
+                            throw;
+                        }
+                    }
+                    return RedirectToAction("Details", new {id = id});
+                }
+                return View(user);
+            }
+            return RedirectToAction(nameof(Login));
+        }
+
+        //GET
+        public async Task<IActionResult> EditContact(int? id) {
+            if(UserIsLoggedIn()) {
+                if (id == null)
+                {
+                    return NotFound();
+                }
+
+                var contact = await _context.Contact
+                    .FirstOrDefaultAsync(m => m.Id == id);
+
+                if (contact == null)
+                {
+                    return NotFound();
+                }
+                
+                return View(contact);
             }
 
-            if (ModelState.IsValid)
-            {
+            return RedirectToAction(nameof(Login));
+        }
+        
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public async Task<IActionResult> EditContact(int id, [Bind("Id,Name,LastName,PhoneNumber,UserId")] Contact contact) {
+            if(UserIsLoggedIn()) {
+                if (id != contact.Id)
+                {
+                    return NotFound();
+                }
+
                 try
                 {
-                    _context.Update(user);
+                    _context.Update(contact);
                     await _context.SaveChangesAsync();
                 }
                 catch (DbUpdateConcurrencyException)
                 {
-                    if (!UserExists(user.Id))
+                    if (!ContactExists(contact.Id))
                     {
                         return NotFound();
                     }
@@ -153,118 +228,90 @@ namespace LoginContacts.Controllers
                         throw;
                     }
                 }
-                return RedirectToAction("Details", new {id = id});
-            }
-            return View(user);
-        }
 
-        //GET
-        public async Task<IActionResult> EditContact(int? id) {
-            if (id == null)
-            {
-                return NotFound();
+                return RedirectToAction("Edit", new {id = contact.UserId});
             }
 
-            var contact = await _context.Contact
-                .FirstOrDefaultAsync(m => m.Id == id);
-
-            if (contact == null)
-            {
-                return NotFound();
-            }
-            
-            return View(contact);
-        }
-        
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> EditContact(int id, [Bind("Id,Name,LastName,PhoneNumber,UserId")] Contact contact) {
-            if (id != contact.Id)
-            {
-                return NotFound();
-            }
-
-            try
-            {
-                _context.Update(contact);
-                await _context.SaveChangesAsync();
-            }
-            catch (DbUpdateConcurrencyException)
-            {
-                if (!ContactExists(contact.Id))
-                {
-                    return NotFound();
-                }
-                else
-                {
-                    throw;
-                }
-            }
-
-            return RedirectToAction("Edit", new {id = contact.UserId});
+            return RedirectToAction(nameof(Login));
         }
 
         // GET
         public async Task<IActionResult> DeleteContact(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
+            if(UserIsLoggedIn()) {
+                if (id == null)
+                {
+                    return NotFound();
+                }
+
+                var contact = await _context.Contact
+                    .FirstOrDefaultAsync(m => m.Id == id);
+                if (contact == null)
+                {
+                    return NotFound();
+                }
+
+                return View(contact);
             }
 
-            var contact = await _context.Contact
-                .FirstOrDefaultAsync(m => m.Id == id);
-            if (contact == null)
-            {
-                return NotFound();
-            }
-
-            return View(contact);
+            return RedirectToAction(nameof(Login));
         }
 
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteContact(int id) {
-            var contact = await _context.Contact
-                .FindAsync(id);
-            
-            if (contact != null)
-            {
-                _context.Contact.Remove(contact);
+            if(UserIsLoggedIn()) {
+                var contact = await _context.Contact
+                    .FindAsync(id);
+                
+                if (contact != null)
+                {
+                    _context.Contact.Remove(contact);
+                }
+
+                await _context.SaveChangesAsync();
+
+                return RedirectToAction("Edit", new {id = contact.UserId});
             }
 
-            await _context.SaveChangesAsync();
-
-            return RedirectToAction("Edit", new {id = contact.UserId});
+            return RedirectToAction(nameof(Login));
         }
         
         public IActionResult CreateContact(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
+            if(UserIsLoggedIn()) {
+                if (id == null)
+                {
+                    return NotFound();
+                }
+
+                Console.WriteLine("Create contact for: " + id);
+
+                ViewData["CurrentUser"] = id;
+                return View();
             }
 
-            Console.WriteLine("Create contact for: " + id);
-
-            ViewData["CurrentUser"] = id;
-            return View();
+            return RedirectToAction(nameof(Login));
         }
 
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> CreateContact([Bind("Id,Name,LastName,PhoneNumber,UserId")] Contact contact)
         {
-            Contact ct = new Contact();
-            ct.Name = contact.Name;
-            ct.LastName = contact.LastName;
-            ct.PhoneNumber = contact.PhoneNumber;
-            ct.UserId = contact.Id;
-            
-            _context.Add(ct);
-            await _context.SaveChangesAsync();
-            
-            return RedirectToAction("Edit", new {id = ct.UserId});
+            if(UserIsLoggedIn()) {
+                Contact ct = new Contact();
+                ct.Name = contact.Name;
+                ct.LastName = contact.LastName;
+                ct.PhoneNumber = contact.PhoneNumber;
+                ct.UserId = contact.Id;
+                
+                _context.Add(ct);
+                await _context.SaveChangesAsync();
+                
+                return RedirectToAction("Edit", new {id = ct.UserId});
+            }
+
+            return RedirectToAction(nameof(Login));
         }
 
         
@@ -272,19 +319,23 @@ namespace LoginContacts.Controllers
         // GET: User/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
+            if(UserIsLoggedIn()) {
+                if (id == null)
+                {
+                    return NotFound();
+                }
+
+                var user = await _context.User
+                    .FirstOrDefaultAsync(m => m.Id == id);
+                if (user == null)
+                {
+                    return NotFound();
+                }
+
+                return View(user);
             }
 
-            var user = await _context.User
-                .FirstOrDefaultAsync(m => m.Id == id);
-            if (user == null)
-            {
-                return NotFound();
-            }
-
-            return View(user);
+            return RedirectToAction(nameof(Login));
         }
 
         // POST: User/Delete/5
@@ -292,15 +343,19 @@ namespace LoginContacts.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            var user = await _context.User
-                .FindAsync(id);
-            if (user != null)
-            {
-                _context.User.Remove(user);
+            if(UserIsLoggedIn()) {
+                var user = await _context.User
+                    .FindAsync(id);
+                if (user != null)
+                {
+                    _context.User.Remove(user);
+                }
+
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
             }
 
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            return RedirectToAction(nameof(Login));
         }
 
         private bool UserExists(int id)
@@ -311,6 +366,10 @@ namespace LoginContacts.Controllers
         private bool ContactExists(int id)
         {
             return _context.Contact.Any(e => e.Id == id);
+        }
+
+        private bool UserIsLoggedIn() {
+            return (HttpContext.Session.GetString("_UserSession") != null);
         }
     }
 }
